Rename and repackage SuggesterService to SuggesterDao and deprecate current API The wrongly-named service was interfacing directly with Solr and exposing implementation details to controllers
